+++ libcoveocds/api.py
@@ -41,51 +41,47 @@ def ocds_json_output(
         lib_cove_ocds_config = LibCoveOCDSConfig()
     if not file_type:
         file_type = get_file_type(file)
-
-    context = {"file_type": file_type}
+    if not json_data and file_type == "json":
+        with open(file, encoding="utf-8") as fp:
+            try:
+                json_data = json.load(fp)
+            except ValueError:
+                raise APIException("The file looks like invalid json")
 
     if file_type == "json":
-        if not json_data:
-            with open(file, encoding="utf-8") as fp:
-                try:
-                    json_data = json.load(fp)
-                except ValueError:
-                    raise APIException("The file looks like invalid json")
-
-        schema_ocds = SchemaOCDS(
+        schema_obj = SchemaOCDS(
             schema_version, json_data, lib_cove_ocds_config=lib_cove_ocds_config, record_pkg=record_pkg
         )
+    else:
+        metatab_schema_url = SchemaOCDS(select_version="1.1", lib_cove_ocds_config=lib_cove_ocds_config).pkg_schema_url
+        metatab_data = get_spreadsheet_meta_data(output_dir, file, metatab_schema_url, file_type=file_type)
+        schema_obj = SchemaOCDS(schema_version, lib_cove_ocds_config=lib_cove_ocds_config, release_data=metatab_data)
 
-        if schema_ocds.invalid_version_data:
-            msg = "\033[1;31mThe schema version in your data is not valid. Accepted values: {}\033[1;m"
-            raise APIException(msg.format(str(list(schema_ocds.version_choices.keys()))))
-        if schema_ocds.extensions:
-            schema_ocds.create_extended_schema_file(output_dir, "")
+    if schema_obj.invalid_version_data:
+        msg = "\033[1;31mThe schema version in your data is not valid. Accepted values: {}\033[1;m"
+        raise APIException(msg.format(str(list(schema_obj.version_choices))))
+    if schema_obj.extensions:
+        schema_obj.create_extended_schema_file(output_dir, "")
+    schema_url = schema_obj.extended_schema_file or schema_obj.schema_url
 
-        url = schema_ocds.extended_schema_file or schema_ocds.schema_url
-
+    context = {"file_type": file_type}
+    if file_type == "json":
         if convert:
             with warnings.catch_warnings():
                 warnings.filterwarnings("ignore")  # flattentool uses UserWarning, so we can't set a specific category
 
                 context.update(
-                    convert_json(output_dir, "", file, lib_cove_ocds_config, schema_url=url, flatten=True, cache=False)
+                    convert_json(
+                        output_dir,
+                        "",
+                        file,
+                        lib_cove_ocds_config,
+                        schema_url=schema_url,
+                        cache=False,
+                        flatten=True,
+                    )
                 )
-
     else:
-        metatab_schema_url = SchemaOCDS(select_version="1.1", lib_cove_ocds_config=lib_cove_ocds_config).pkg_schema_url
-        metatab_data = get_spreadsheet_meta_data(output_dir, file, metatab_schema_url, file_type=file_type)
-        schema_ocds = SchemaOCDS(schema_version, release_data=metatab_data, lib_cove_ocds_config=lib_cove_ocds_config)
-
-        if schema_ocds.invalid_version_data:
-            msg = "\033[1;31mThe schema version in your data is not valid. Accepted values: {}\033[1;m"
-            raise APIException(msg.format(str(list(schema_ocds.version_choices.keys()))))
-        if schema_ocds.extensions:
-            schema_ocds.create_extended_schema_file(output_dir, "")
-
-        url = schema_ocds.extended_schema_file or schema_ocds.schema_url
-        pkg_url = schema_ocds.pkg_schema_url
-
         context.update(
             convert_spreadsheet(
                 output_dir,
@@ -93,9 +89,9 @@ def ocds_json_output(
                 file,
                 file_type,
                 lib_cove_ocds_config,
-                schema_url=url,
-                pkg_schema_url=pkg_url,
+                schema_url=schema_url,
                 cache=False,
+                pkg_schema_url=schema_obj.pkg_schema_url,
             )
         )
 
@@ -105,7 +101,7 @@ def ocds_json_output(
     context = context_api_transform(
         # This `cache` writes the results to a file, which is only relevant in the context of repetitive web requests.
         common_checks_ocds(
-            context, output_dir, json_data, schema_ocds, api=True, cache=False, skip_aggregates=skip_aggregates
+            context, output_dir, json_data, schema_obj, api=True, cache=False, skip_aggregates=skip_aggregates
         )
     )