+++ CHANGELOG.md
@@ -9,12 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- Improve performance in API context:
-  - Add `skip_aggregates` keyword argument to `common_checks_ocds` and `ocds_json_output` to skip "releases_aggregates" and "records_aggregates" (default `False`).
-  - Skip Markdown formatting, HTML escaping and adding fields to validation errors that are then ignored by ``context_api_transform()``.
-  - Skip adding the field description and reference URL of the `ocid` field for incorrectly formatted OCID prefixes.
+**BREAKING CHANGES:**
+
+- `libcoveocds.lib.common_checks`: Rename `get_bad_ocds_prefixes` to `get_bad_ocid_prefixes`.
+
+Other changes:
+
+- Improve performance in API context.
+  - Add `skip_aggregates` keyword argument to `common_checks_ocds` and `ocds_json_output` (default `False`).
+  - Skip the schema description and reference URL for OCID prefix conformance errors.
+  - Skip the formatted message, schema title, schema description and reference URL for validation errors.
   - Skip sorting the JSON locations of additional checks.
   - Improve ``context_api_transform()`` performance.
+- Use ocdsextensionregistry to merge extensions.
+  - Empty extension URLs no longer produce an error message.
 - flattentool is optional.
 - Drop support for Python 3.7.